Day 1 completed
Code is not the greatest but it works without issue. Node is a bit strange to create simple scripts. I might need to do more reading to see if I can do this with es6 and more advanced tools. This way I can get the code to be cleaner and easier to read.